refactor: 예약 생성 로직 메서드 분리
